requirements.txt
python-dotenv
webdriver-manager
selenium>=4.0.0
pytest>=7.0.0
pytest-html>=3.2.0
openpyxl>=3.1.0
Pillow>=10.0.0